+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/ShareButton/utils.tsx
@@ -1,3 +1,4 @@
+import { NextRouter } from "next/router";
 import { Dispatch } from "react";
 import { TissueMetadataQueryResponse } from "src/common/queries/cellGuide";
 import { isSSR } from "src/common/utils/isSSR";
@@ -76,17 +77,19 @@ const stripEmptyFilters = (
 };
 
 export const loadStateFromQueryParams = ({
-  params,
-  selectedFilters,
-  dispatch,
-  tissues,
   cellTypesByName,
+  dispatch,
+  params,
+  router,
+  selectedFilters,
+  tissues,
 }: {
   params: URLSearchParams;
   selectedFilters: State["selectedFilters"];
   dispatch: Dispatch<PayloadAction<LoadStateFromURLPayload>>;
   tissues?: TissueMetadataQueryResponse;
   cellTypesByName: { [name: string]: CellType };
+  router: NextRouter;
 }): LoadStateFromURLPayload | null => {
   if (isSSR()) return null;
 
@@ -153,7 +156,7 @@ export const loadStateFromQueryParams = ({
    * (thuang): Please makes sure we only remove params AFTER pushing all params
    * to `paramsToRemove`
    */
-  removeParams(paramsToRemove);
+  removeParams({ params: paramsToRemove, router });
 
   const payload = {
     compare: newCompare,
