+++ frontend/src/views/WheresMyGeneV2/components/HeatMap/types.ts
@@ -1,0 +1,40 @@
+import { Dispatch, SetStateAction } from "react";
+
+import { CellTypeRow, OntologyTerm } from "src/common/queries/wheresMyGene";
+import { State } from "src/views/WheresMyGene/common/store";
+
+import {
+  ChartProps,
+  GeneExpressionSummary,
+  SORT_BY,
+  Tissue,
+} from "src/views/WheresMyGene/common/types";
+
+export interface Props {
+  className?: string;
+  cellTypes: { [tissue: Tissue]: CellTypeRow[] };
+  genes: State["selectedGenes"];
+  selectedGeneExpressionSummariesByTissueName: {
+    [tissueName: string]: GeneExpressionSummary[];
+  };
+  scaledMeanExpressionMax: number;
+  scaledMeanExpressionMin: number;
+  isLoadingAPI: boolean;
+  isScaled: boolean;
+  cellTypeSortBy: SORT_BY;
+  geneSortBy: SORT_BY;
+  echartsRendererMode: "svg" | "canvas";
+  setAllChartProps: Dispatch<
+    SetStateAction<{
+      [tissue: string]: ChartProps;
+    }>
+  >;
+  allChartProps: { [tissue: string]: ChartProps };
+  tissuesByName: { [name: string]: OntologyTerm };
+  setTissuesByName: Dispatch<
+    SetStateAction<{
+      [name: string]: OntologyTerm;
+    }>
+  >;
+  sidebarWidth: number;
+}
